task 2/3: Extract TokenInput/TokenChip into components/TokenInput.kt with Src/Dst wrappers and shared loadingPlaceholder, plus components/PercentagePicker.kt (#4836)
