+++ app/src/main/java/com/vultisig/wallet/ui/screens/swap/SwapScreen.kt
@@ -1,6 +1,5 @@
 package com.vultisig.wallet.ui.screens.swap
 
-import android.annotation.SuppressLint
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.core.Animatable
@@ -19,18 +18,14 @@ import androidx.compose.animation.slideOutVertically
 import androidx.compose.animation.togetherWith
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.gestures.detectDragGesturesAfterLongPress
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.interaction.collectIsFocusedAsState
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.IntrinsicSize
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
@@ -45,7 +40,6 @@ import androidx.compose.foundation.text.input.TextFieldLineLimits
 import androidx.compose.foundation.text.input.TextFieldState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -59,17 +53,12 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.rotate
 import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.graphics.StrokeCap
 import androidx.compose.ui.graphics.graphicsLayer
-import androidx.compose.ui.input.pointer.PointerInputChange
-import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.layout.boundsInParent
 import androidx.compose.ui.layout.onGloballyPositioned
-import androidx.compose.ui.layout.positionInWindow
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.painterResource
@@ -88,28 +77,26 @@ import androidx.navigation.NavHostController
 import com.vultisig.wallet.R
 import com.vultisig.wallet.data.models.SwapQuote.Companion.expiredAfter
 import com.vultisig.wallet.data.utils.timerFlow
-import com.vultisig.wallet.ui.components.TokenAndChainLogo
 import com.vultisig.wallet.ui.components.UiIcon
 import com.vultisig.wallet.ui.components.UiSpacer
 import com.vultisig.wallet.ui.components.buttons.VsButton
 import com.vultisig.wallet.ui.components.buttons.VsButtonState
 import com.vultisig.wallet.ui.components.buttons.VsButtonVariant
 import com.vultisig.wallet.ui.components.inputs.VsBasicTextField
-import com.vultisig.wallet.ui.components.library.UiPlaceholderLoader
 import com.vultisig.wallet.ui.components.library.form.FormDetails2
 import com.vultisig.wallet.ui.components.rememberKeyboardVisibilityAsState
-import com.vultisig.wallet.ui.components.selectors.ChainSelector
-import com.vultisig.wallet.ui.components.util.CutoutPosition
-import com.vultisig.wallet.ui.components.util.RoundedWithCutoutShape
 import com.vultisig.wallet.ui.components.v2.fastselection.contentWithFastSelection
 import com.vultisig.wallet.ui.components.v2.scaffold.V2Scaffold
 import com.vultisig.wallet.ui.components.v2.utils.toPx
-import com.vultisig.wallet.ui.models.send.TokenBalanceUiModel
 import com.vultisig.wallet.ui.models.swap.SwapFormUiModel
 import com.vultisig.wallet.ui.models.swap.SwapFormViewModel
 import com.vultisig.wallet.ui.navigation.Route
 import com.vultisig.wallet.ui.screens.settings.TierType
+import com.vultisig.wallet.ui.screens.swap.components.DstTokenInput
 import com.vultisig.wallet.ui.screens.swap.components.HintBox
+import com.vultisig.wallet.ui.screens.swap.components.PercentagePicker
+import com.vultisig.wallet.ui.screens.swap.components.SrcTokenInput
+import com.vultisig.wallet.ui.screens.swap.components.loadingPlaceholder
 import com.vultisig.wallet.ui.theme.Theme
 import com.vultisig.wallet.ui.utils.asString
 import java.util.Locale
@@ -239,23 +226,15 @@ internal fun SwapScreen(
                                     )
                                 }
 
-                                TokenInput(
+                                SrcTokenInput(
                                     isLoading = state.isLoading,
                                     title = stringResource(R.string.swap_form_from_title),
                                     selectedToken = state.selectedSrcToken,
                                     fiatValue = state.srcFiatValue,
+                                    space = space,
                                     onSelectNetworkClick = onSelectSrcNetworkClick,
                                     onSelectTokenClick = onSelectSrcToken,
-                                    chainTestTag = "SwapFormScreen.fromChain",
-                                    tokenTestTag = "SwapFormScreen.fromToken",
-                                    shape =
-                                        RoundedWithCutoutShape(
-                                            cutoutPosition = CutoutPosition.Bottom,
-                                            cutoutOffsetY = -space / 2,
-                                            cutoutRadius = 28.dp,
-                                            onCircleBoundsChanged = { topCenter = it },
-                                        ),
-                                    focused = true,
+                                    onCircleBoundsChanged = { topCenter = it },
                                     onDrag = onDrag,
                                     onDragEnd = onDragEnd,
                                     onDragCancel = onDragCancel,
@@ -352,28 +331,20 @@ internal fun SwapScreen(
                             val dstHasValue =
                                 state.quoteDisplay.estimatedDstTokenValue.isNotBlank() &&
                                     state.quoteDisplay.estimatedDstTokenValue != "0"
-                            TokenInput(
+                            DstTokenInput(
                                 title = stringResource(R.string.swap_form_dst_token_title),
                                 isLoading = state.isLoading && !dstHasValue,
                                 selectedToken = state.selectedDstToken,
                                 fiatValue = state.quoteDisplay.estimatedDstFiatValue,
+                                space = space,
                                 onSelectNetworkClick = onSelectDstNetworkClick,
                                 onSelectTokenClick = onSelectDstToken,
-                                chainTestTag = "SwapFormScreen.toChain",
-                                tokenTestTag = "SwapFormScreen.toToken",
-                                shape =
-                                    RoundedWithCutoutShape(
-                                        cutoutPosition = CutoutPosition.Top,
-                                        cutoutOffsetY = -space / 2,
-                                        cutoutRadius = 28.dp,
-                                        onCircleBoundsChanged = { bottomCenter = it },
-                                    ),
+                                onCircleBoundsChanged = { bottomCenter = it },
                                 onDrag = onDrag,
                                 onDragEnd = onDragEnd,
                                 onDragCancel = onDragCancel,
                                 onDragStart = onDragStart,
                                 onLongPressStarted = onDstLongPressStarted,
-                                focused = false,
                                 textFieldContent = {
                                     Text(
                                         text = state.quoteDisplay.estimatedDstTokenValue,
@@ -397,13 +368,12 @@ internal fun SwapScreen(
                             verticalArrangement = Arrangement.spacedBy(10.dp),
                             modifier = Modifier.padding(horizontal = 8.dp),
                         ) {
-                            val placeHolderModifier = Modifier.height(16.dp).width(80.dp)
                             FormDetails2(
                                 title = stringResource(R.string.swap_screen_provider_title),
                                 value = state.quoteDisplay.provider.asString(),
                                 placeholder =
                                     if (state.isLoading) {
-                                        { UiPlaceholderLoader(placeHolderModifier) }
+                                        { loadingPlaceholder() }
                                     } else null,
                             )
 
@@ -415,7 +385,7 @@ internal fun SwapScreen(
                                 title = stringResource(R.string.swap_form_total_fees_title),
                                 valueComposable =
                                     if (state.isLoading) {
-                                        { UiPlaceholderLoader(placeHolderModifier) }
+                                        { loadingPlaceholder() }
                                     } else {
                                         {
                                             Row {
@@ -494,7 +464,7 @@ internal fun SwapScreen(
                                                 },
                                             placeholder =
                                                 if (state.isLoading) {
-                                                    { UiPlaceholderLoader(placeHolderModifier) }
+                                                    { loadingPlaceholder() }
                                                 } else null,
                                         )
 
@@ -514,7 +484,7 @@ internal fun SwapScreen(
                                             value = state.feeBreakdown.fee,
                                             placeholder =
                                                 if (state.isLoading) {
-                                                    { UiPlaceholderLoader(placeHolderModifier) }
+                                                    { loadingPlaceholder() }
                                                 } else null,
                                         )
 
@@ -624,32 +594,10 @@ internal fun SwapScreen(
                 },
             ) { showPercentagePicker ->
                 if (showPercentagePicker) {
-                    Column {
-                        HorizontalDivider(thickness = 1.dp, color = Theme.v2.colors.border.light)
-
-                        Row(
-                            horizontalArrangement = Arrangement.spacedBy(8.dp),
-                            modifier =
-                                Modifier.fillMaxWidth()
-                                    .background(color = Theme.v2.colors.backgrounds.secondary)
-                                    .padding(vertical = 12.dp, horizontal = 8.dp),
-                        ) {
-                            PercentageItem(
-                                title = "25%",
-                                onClick = { onSelectSrcPercentage(0.25f) },
-                            )
-                            PercentageItem(title = "50%", onClick = { onSelectSrcPercentage(0.5f) })
-                            PercentageItem(
-                                title = "75%",
-                                onClick = { onSelectSrcPercentage(0.75f) },
-                            )
-                            if (state.enableMaxAmount)
-                                PercentageItem(
-                                    title = "MAX",
-                                    onClick = { onSelectSrcPercentage(1f) },
-                                )
-                        }
-                    }
+                    PercentagePicker(
+                        enableMaxAmount = state.enableMaxAmount,
+                        onSelectSrcPercentage = onSelectSrcPercentage,
+                    )
                 } else {
                     VsButton(
                         label =
@@ -677,24 +625,6 @@ internal fun SwapScreen(
                 }
             }
         },
-    )
-}
-
-@Composable
-private fun RowScope.PercentageItem(title: String, onClick: () -> Unit) {
-    Text(
-        text = title,
-        style = Theme.brockmann.supplementary.caption,
-        color = Theme.v2.colors.text.primary,
-        textAlign = TextAlign.Center,
-        modifier =
-            Modifier.clickable(onClick = onClick)
-                .background(
-                    color = Theme.v2.colors.backgrounds.tertiary_2,
-                    shape = RoundedCornerShape(99.dp),
-                )
-                .padding(all = 8.dp)
-                .weight(1f),
     )
 }
 
@@ -749,199 +679,6 @@ private fun VultDiscountTier(vultBpsDiscount: Int, tierType: TierType?) {
             color = Theme.v2.colors.text.tertiary,
             style = Theme.brockmann.supplementary.caption,
         )
-    }
-}
-
-@Composable
-private fun TokenInput(
-    isLoading: Boolean,
-    title: String,
-    selectedToken: TokenBalanceUiModel?,
-    fiatValue: String,
-    onSelectNetworkClick: () -> Unit,
-    onSelectTokenClick: () -> Unit,
-    shape: Shape,
-    focused: Boolean,
-    modifier: Modifier = Modifier,
-    chainTestTag: String? = null,
-    tokenTestTag: String? = null,
-    @SuppressLint("ComposableLambdaParameterNaming") onDragStart: (Offset) -> Unit = {},
-    onDrag: (Offset) -> Unit = {},
-    onDragEnd: () -> Unit = {},
-    onDragCancel: () -> Unit = {},
-    onLongPressStarted: (Offset) -> Unit = {},
-    textFieldContent: @Composable ColumnScope.() -> Unit,
-) {
-    Column(
-        verticalArrangement = Arrangement.spacedBy(16.dp),
-        modifier =
-            modifier
-                .background(
-                    color =
-                        if (focused) Theme.v2.colors.backgrounds.secondary
-                        else Theme.v2.colors.backgrounds.disabled,
-                    shape = shape,
-                )
-                .border(width = 1.dp, color = Theme.v2.colors.border.light, shape = shape)
-                .clip(shape)
-                .padding(all = 16.dp),
-    ) {
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.SpaceBetween,
-            modifier = Modifier.fillMaxWidth(),
-        ) {
-            val selectedChain = selectedToken?.model?.address?.chain
-
-            if (selectedChain != null) {
-                Box(
-                    modifier =
-                        if (chainTestTag != null) Modifier.testTag(chainTestTag) else Modifier
-                ) {
-                    ChainSelector(
-                        title = title,
-                        chain = selectedChain,
-                        onClick = onSelectNetworkClick,
-                        onDragStart = onDragStart,
-                        onDragCancel = onDragCancel,
-                        onDragEnd = onDragEnd,
-                        onDrag = onDrag,
-                        onLongPressStarted = onLongPressStarted,
-                    )
-                }
-            }
-
-            Text(
-                text = selectedToken?.let { "${it.balance} ${it.title}" } ?: "",
-                style = Theme.brockmann.supplementary.caption,
-                color = Theme.v2.colors.text.tertiary,
-                textAlign = TextAlign.End,
-            )
-        }
-
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.SpaceBetween,
-            modifier = Modifier.fillMaxWidth(),
-        ) {
-            Box(modifier = if (tokenTestTag != null) Modifier.testTag(tokenTestTag) else Modifier) {
-                TokenChip(selectedToken = selectedToken, onSelectTokenClick = onSelectTokenClick)
-            }
-
-            Column(
-                verticalArrangement = Arrangement.spacedBy(6.dp),
-                horizontalAlignment = Alignment.End,
-            ) {
-                if (isLoading && title == stringResource(R.string.swap_form_dst_token_title)) {
-                    UiPlaceholderLoader(modifier = Modifier.height(24.dp).width(150.dp))
-                } else {
-                    textFieldContent()
-                }
-
-                if (isLoading) {
-                    UiPlaceholderLoader(modifier = Modifier.height(16.dp).width(80.dp))
-                } else {
-                    Text(
-                        text = fiatValue,
-                        style = Theme.brockmann.supplementary.caption,
-                        color = Theme.v2.colors.text.tertiary,
-                        textAlign = TextAlign.End,
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-internal fun TokenChip(
-    selectedToken: TokenBalanceUiModel?,
-    onSelectTokenClick: () -> Unit,
-    onDragStart: (Offset) -> Unit = {},
-    onDrag: (Offset) -> Unit = {},
-    onDragEnd: () -> Unit = {},
-    onDragCancel: () -> Unit = {},
-    onLongPressStarted: (Offset) -> Unit = {},
-) {
-
-    var fieldPosition by remember { mutableStateOf(Offset.Zero) }
-
-    Row(
-        verticalAlignment = Alignment.CenterVertically,
-        modifier =
-            Modifier.clickable(onClick = onSelectTokenClick)
-                .onGloballyPositioned { coordinates ->
-                    fieldPosition = coordinates.positionInWindow()
-                }
-                .pointerInput(Unit) {
-                    detectDragGesturesAfterLongPress(
-                        onDragStart = { offset ->
-                            val screenPosition =
-                                Offset(
-                                    x = fieldPosition.x + offset.x,
-                                    y = fieldPosition.y + offset.y,
-                                )
-                            onDragStart(screenPosition)
-                            onLongPressStarted(screenPosition)
-                        },
-                        onDrag = { change: PointerInputChange, _ ->
-                            val localPos = change.position
-                            val screenPos =
-                                Offset(
-                                    x = fieldPosition.x + localPos.x,
-                                    y = fieldPosition.y + localPos.y,
-                                )
-                            onDrag(screenPos)
-                            change.consume()
-                        },
-                        onDragEnd = onDragEnd,
-                        onDragCancel = onDragCancel,
-                    )
-                }
-                .background(
-                    color = Theme.v2.colors.backgrounds.tertiary_2,
-                    shape = RoundedCornerShape(99.dp),
-                )
-                .padding(all = 6.dp),
-    ) {
-        TokenAndChainLogo(
-            tokenLogo = selectedToken?.tokenLogo ?: "",
-            tokenTicker = selectedToken?.title ?: "",
-            chainLogo =
-                selectedToken?.chainLogo.takeIf {
-                    selectedToken?.isNativeToken == false || selectedToken?.isLayer2 == true
-                },
-            chainLogoSize = 16.dp,
-            tokenLogoSize = 36.dp,
-        )
-
-        UiSpacer(8.dp)
-
-        Column {
-            Text(
-                text = selectedToken?.title ?: "",
-                style = Theme.brockmann.supplementary.caption,
-                color = Theme.v2.colors.text.primary,
-            )
-
-            if (selectedToken?.isNativeToken == true) {
-                Text(
-                    text = stringResource(R.string.swap_form_native),
-                    style = Theme.brockmann.supplementary.captionSmall,
-                    color = Theme.v2.colors.text.tertiary,
-                )
-            }
-        }
-
-        UiSpacer(4.dp)
-
-        UiIcon(
-            drawableResId = R.drawable.ic_chevron_right_small,
-            size = 20.dp,
-            tint = Theme.v2.colors.text.primary,
-        )
-
-        UiSpacer(6.dp)
     }
 }
